requirements.txt
# Основные зависимости
python-telegram-bot[job-queue]>=22.0
python-dotenv>=1.0.0

# HTTP клиенты (асинхронные)
httpx>=0.27.0
aiohttp>=3.9.0

# База данных
sqlalchemy>=2.0.0
aiosqlite>=0.19.0  # Для асинхронного SQLite
asyncpg>=0.29.0  # PostgreSQL для высокой нагрузки (~80k пользователей)
alembic>=1.13.0  # Миграции БД

# Валидация данных и конфигурация
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Retry механизм
tenacity>=8.2.0

# Логирование (опционально)
structlog>=23.2.0

# Redis + очереди задач
redis>=5.0.0
hiredis>=2.2.0
taskiq>=0.10.0
taskiq-redis>=0.1.0

# Тестирование
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Дополнительные утилиты
APScheduler>=3.10.0

# Наблюдение за файлами (для автоматического бэкапа)
watchdog>=3.0.0

# Админ-панель
streamlit>=1.28.0
pandas>=2.0.0

# Ruff (линтер)
ruff>=0.1.0

# Observability (Prometheus)
prometheus_client>=0.19.0

# RAG: PDF и векторная БД
pypdf>=4.0.0
chromadb>=0.4.0
